First commit: minimalistic functional version of my idea
Here is what I have in mind: to make a board that is great for many
people like the Kyria, compromises must be taken.
For instance, footprints must be versatile to allow for various types
of switches, socketed or non-socketed etc. Supporting a board like the
pro-micro or elite-c allows for versatility while keeping the pcb design
a bit easier. Supporting a symmetrical layout allows for a simpler
printing process for the pcb.

With versatile footprints, there can be unnecessary holes for any given
configuration that make the board less aesthaetically pleasing and
harder to solder on. Some switches may also end up in strange
orientations.
With processor boards, thickness is added to the keyboard and more space
is taken.
With symmetrical layouts, even more unnecessary holes are introduced,
there is more room for mistakes. The cost of each board goes up as the
boards use 2 processors where one should be able to do the job. The
firmware is more complicated. Flashing is a bit more involved.

These are just a few examples, but the point is: by tailoring a board to
a specific usecase, I could get a board that is excellent at that
purpose and more elegant for this usecase than a more generic board.
It may be interesting to address different usecases too, but that can be
done by creating a new pcb design based on this one as opposed to trying
to provide everything on one board.

The ferris board is a no-fuss low-profile split keyboard with 36 keys
laid out exactly like the 36 keys that I consider to be the most
accessible on the kyria. It is meant to be used without a case, simply
dressed with choc switches and keycaps. All the components are smd
mounted on the top of the pcb for eye-candy and ease of use without a
bottom plate to protect the components. The left hand is the master and
hosts an atmega32u4 (the board is very sparse in features so I don't
reckon it needs any more power). It connects to the laptop through a usb
type-c port (I am so over non-type-c connectors!). Both halves are
connected with a TRRS jack cable. The right hand side is simply a
continuation of the switch matrix and uses an MCP-23017 microchip to
essentially unpack the signal from the 2 data pins coming from the TRRS
jack into more GPIO pins than we need. There is no backlighting for
simplicity, which allows to steal the per key backlighting location for
the diodes for a very bare aesthaetic once assembled. The traces are
routed in a way that tries to expose the functioning of the pcb, with
traces for the matrix's rows and most of the components prominently
featured in the front while the back mostly exposes the columns of the
matrix.

This is my very first pcb design of any kind, so I will seek feedback
and I expect to find issues in the design that need addressing before
printing any samples.

I took inspiraton from a number of boards, which I would like to
credit:
The Kyria, my first split keyboard ever. I re-used its layout as I find
it very comfortable. The crkbd for the idea of proposing a variety of
flavours for a board instead of doing too much in a single body. The
gergoplex for its focus on simplicity. The arch 36 for the idea of using
the Kyria layout for making a simpler board.

Note that this is still only a first draft and not ready for prototyping
yet. I plan to add some artwork in the silkscreen and possibly in a
custom pad, but I will have to learn how to do these things first. I
also plan to touch up the tracing a bit more. In particular, I plan to
add a vcc and a ground pane around the mpc-23017 for symmetry with the
right hand usb-c area and to loose some wire clutter.