Add Ferris copper pad and various cleanups
Add Ferris copper pad
Add front and back masks behind the right hand's usb area for
consistency with the left side and to clean up a couple of traces.
Reroute the entire highway between the atmega and the usb c area on the
left hand.
Add the cleaned up switches footprints to ferris.pretty and use these in
the board.
Add a render of the front of the pcb for the readme.