Create MX layout for the Ferris, by Melinda Birkenstock
This change was contributed by Melinda Birkenstock and is the first step
towards making Ferris-MX happen: it updates all the footprints to MX
only-non symetrical footprints, in the spirit of the "do one thing well"
philosophy the Ferris strives for.
The traces are neatly routed in this configuration with nice and
straight columns.