Replace meson with a simple python script and ninja
Already with my usage, I was hitting many "opinions" from meson that
were forcing me to write a lot of boilerplate.
meson is designed to build very conforming C++ projects and not much
else, so it wasn't the right tool for the job for something custom and
very simple.

Replace it with ~100 lines of python for now to generate a ninja file
that can be committed.

The user workflow is as simple as can be:
```
ninja
```
The build dir will contain all outputs.

For adding variants or artifacts,
* Edit `configure.py`
* `./configure.py`
* `git add build.ninja`