Use docker for generating artifacts
Generating all artifacts should now be as simple as running:
```
./ninja.sh
```
with the only prerequisite being to have docker installed and
configured.

This solution is not perfect at the moment as the artifacts produced are
owned by the root user, which is less than ideal; but the outcome is
more reliable than it has ever been.