0.2/bling: Add human readable "Part Name" field
Use short human readable names to describe components, such as "MCU"
instead of "STM32F072CBT6".

This allows us to be quite precise in the "value" field while displaying
an easy to understand description of each component in the generated
interactive Bill of Material.

We can then use the same language on the BOM cards for a smooth user
experience of assembling one's Ferris.