Make Bling mid-profile case more robust
I plan to print the top and bottom plates in 2mm acrylic, so thin parts
will be quite fragile.
Use the new layout of the components in the PCB to tweak the case with a
little more rigidity.

Also rename the files so they are a bit more readable.

Changes:
* Lip is now a closed shape around the USB area (more solid)
* Lip is thicker: 5 mm in thickest point instead of 4mm previously
* Cutouts for components are much tighter to the components
* Remove the non-puck switch plate as it is better to give people the
  option to upgrade later if they so desire.